refactor comment functions into comment files from review files
